Split up init part of NASA9 unit test
We can reuse this for other tests, so let's split it off
into its own class that we can subclass with both the current
NASA9CurveFit test and other future tests.
