+++ test/standard_unit/nasa9_curve_fit_test.C
@@ -31,44 +31,27 @@
 // C++
 #include <limits>
 
-// CppUnit
-#include <cppunit/extensions/HelperMacros.h>
-#include <cppunit/TestCase.h>
-
 // Antioch
 #include "antioch/nasa9_curve_fit.h"
 #include "antioch/temp_cache.h"
+#include "nasa9_thermo_test_base.h"
 
 namespace AntiochTesting
 {
   template<typename Scalar>
-  class NASA9CurveFitTest : public CppUnit::TestCase
+  class NASA9CurveFitTest : public NASA9ThermoTestBase<Scalar>
   {
   public:
-
-    void setUp()
-    {
-      this->init_N2_coeffs_200_1000();
-      this->init_N2_coeffs_1000_6000();
-      this->init_N2_coeffs_6000_20000();
-      this->init_all_N2_coeffs();
-
-      this->init_NO2_coeffs_200_1000();
-      this->init_NO2_coeffs_1000_6000();
-      this->init_all_NO2_coeffs();
-    }
-
-    void tearDown(){}
 
     void test_nasa9_default_temp_intervals()
     {
       {
-        Antioch::NASA9CurveFit<Scalar> curve_fit( _all_N2_coeffs );
+        Antioch::NASA9CurveFit<Scalar> curve_fit( this->_all_N2_coeffs );
         this->three_interval_test(curve_fit);
       }
 
       {
-        Antioch::NASA9CurveFit<Scalar> curve_fit( _all_NO2_coeffs );
+        Antioch::NASA9CurveFit<Scalar> curve_fit( this->_all_NO2_coeffs );
         this->two_interval_test(curve_fit);
       }
 
@@ -83,7 +66,7 @@ namespace AntiochTesting
         temp[2] = 6000;
         temp[3] = 20000;
 
-        Antioch::NASA9CurveFit<Scalar> curve_fit( _all_N2_coeffs, temp );
+        Antioch::NASA9CurveFit<Scalar> curve_fit( this->_all_N2_coeffs, temp );
         this->three_interval_test(curve_fit);
       }
 
@@ -93,7 +76,7 @@ namespace AntiochTesting
         temp[1] = 1000;
         temp[2] = 6000;
 
-        Antioch::NASA9CurveFit<Scalar> curve_fit( _all_NO2_coeffs, temp );
+        Antioch::NASA9CurveFit<Scalar> curve_fit( this->_all_NO2_coeffs, temp );
         this->two_interval_test(curve_fit);
       }
 
@@ -102,32 +85,32 @@ namespace AntiochTesting
     void three_interval_test( const Antioch::NASA9CurveFit<Scalar>& curve_fit )
     {
       Scalar T = 302.0;
-      this->test_cp( T, _N2_coeffs_200_1000, curve_fit );
-      this->test_h( T, _N2_coeffs_200_1000, curve_fit );
-      this->test_s( T, _N2_coeffs_200_1000, curve_fit );
+      this->test_cp( T, this->_N2_coeffs_200_1000, curve_fit );
+      this->test_h( T, this->_N2_coeffs_200_1000, curve_fit );
+      this->test_s( T, this->_N2_coeffs_200_1000, curve_fit );
 
       T = 3002.0;
-      this->test_cp( T, _N2_coeffs_1000_6000, curve_fit );
-      this->test_h( T, _N2_coeffs_1000_6000, curve_fit );
-      this->test_s( T, _N2_coeffs_1000_6000, curve_fit );
+      this->test_cp( T, this->_N2_coeffs_1000_6000, curve_fit );
+      this->test_h( T, this->_N2_coeffs_1000_6000, curve_fit );
+      this->test_s( T, this->_N2_coeffs_1000_6000, curve_fit );
 
       T = 7002.0;
-      this->test_cp( T, _N2_coeffs_6000_20000, curve_fit );
-      this->test_h( T, _N2_coeffs_6000_20000, curve_fit );
-      this->test_s( T, _N2_coeffs_6000_20000, curve_fit );
+      this->test_cp( T, this->_N2_coeffs_6000_20000, curve_fit );
+      this->test_h( T, this->_N2_coeffs_6000_20000, curve_fit );
+      this->test_s( T, this->_N2_coeffs_6000_20000, curve_fit );
     }
 
     void two_interval_test( const Antioch::NASA9CurveFit<Scalar>& curve_fit )
     {
       Scalar T = 302.0;
-      this->test_cp( T, _NO2_coeffs_200_1000, curve_fit );
-      this->test_h( T, _NO2_coeffs_200_1000, curve_fit );
-      this->test_s( T, _NO2_coeffs_200_1000, curve_fit );
+      this->test_cp( T, this->_NO2_coeffs_200_1000, curve_fit );
+      this->test_h( T, this->_NO2_coeffs_200_1000, curve_fit );
+      this->test_s( T, this->_NO2_coeffs_200_1000, curve_fit );
 
       T = 3002.0;
-      this->test_cp( T, _NO2_coeffs_1000_6000, curve_fit );
-      this->test_h( T, _NO2_coeffs_1000_6000, curve_fit );
-      this->test_s( T, _NO2_coeffs_1000_6000, curve_fit );
+      this->test_cp( T, this->_NO2_coeffs_1000_6000, curve_fit );
+      this->test_h( T, this->_NO2_coeffs_1000_6000, curve_fit );
+      this->test_s( T, this->_NO2_coeffs_1000_6000, curve_fit );
     }
 
     void test_cp( Scalar T,
@@ -208,113 +191,6 @@ namespace AntiochTesting
         + a5*(T*T*T)/3.0L + a6*(T*T*T*T)/4.0L + a8;
     }
 
-    void init_N2_coeffs_200_1000()
-    {
-      _N2_coeffs_200_1000.resize(9);
-      _N2_coeffs_200_1000[0] =  2.21037122e+04L;
-      _N2_coeffs_200_1000[1] = -3.81846145e+02L;
-      _N2_coeffs_200_1000[2] =  6.08273815e+00L;
-      _N2_coeffs_200_1000[3] = -8.53091381e-03L;
-      _N2_coeffs_200_1000[4] =  1.38464610e-05L;
-      _N2_coeffs_200_1000[5] = -9.62579293e-09L;
-      _N2_coeffs_200_1000[6] =  2.51970560e-12L;
-      _N2_coeffs_200_1000[7] =  7.10845911e+02L;
-      _N2_coeffs_200_1000[8] = -1.07600320e+01L;
-    }
-
-    void init_N2_coeffs_1000_6000()
-    {
-      _N2_coeffs_1000_6000.resize(9);
-      _N2_coeffs_1000_6000[0] =  5.87709908e+05L;
-      _N2_coeffs_1000_6000[1] = -2.23924255e+03L;
-      _N2_coeffs_1000_6000[2] =  6.06694267e+00L;
-      _N2_coeffs_1000_6000[3] = -6.13965296e-04L;
-      _N2_coeffs_1000_6000[4] =  1.49179819e-07L;
-      _N2_coeffs_1000_6000[5] = -1.92309442e-11L;
-      _N2_coeffs_1000_6000[6] =  1.06194871e-15L;
-      _N2_coeffs_1000_6000[7] =  1.28320618e+04L;
-      _N2_coeffs_1000_6000[8] = -1.58663484e+01L;
-    }
-
-    void init_N2_coeffs_6000_20000()
-    {
-      _N2_coeffs_6000_20000.resize(9);
-      _N2_coeffs_6000_20000[0] =  8.30971200e+08L;
-      _N2_coeffs_6000_20000[1] = -6.42048187e+05L;
-      _N2_coeffs_6000_20000[2] =  2.02020507e+02L;
-      _N2_coeffs_6000_20000[3] = -3.06501961e-02L;
-      _N2_coeffs_6000_20000[4] =  2.48685558e-06L;
-      _N2_coeffs_6000_20000[5] = -9.70579208e-11L;
-      _N2_coeffs_6000_20000[6] =  1.43751673e-15L;
-      _N2_coeffs_6000_20000[7] =  4.93850663e+06L;
-      _N2_coeffs_6000_20000[8] = -1.67204791e+03L;
-    }
-
-    void init_all_N2_coeffs()
-    {
-      _all_N2_coeffs.insert(  _all_N2_coeffs.end(),
-                              _N2_coeffs_200_1000.begin(),
-                              _N2_coeffs_200_1000.end() );
-
-      _all_N2_coeffs.insert(  _all_N2_coeffs.end(),
-                              _N2_coeffs_1000_6000.begin(),
-                              _N2_coeffs_1000_6000.end() );
-
-      _all_N2_coeffs.insert(  _all_N2_coeffs.end(),
-                              _N2_coeffs_6000_20000.begin(),
-                              _N2_coeffs_6000_20000.end() );
-    }
-
-    void init_NO2_coeffs_200_1000()
-    {
-      _NO2_coeffs_200_1000.resize(9);
-      _NO2_coeffs_200_1000[0] = -5.64204584e+04L;
-      _NO2_coeffs_200_1000[1] =  9.63309734e+02L;
-      _NO2_coeffs_200_1000[2] = -2.43451851e+00L;
-      _NO2_coeffs_200_1000[3] =  1.92776414e-02L;
-      _NO2_coeffs_200_1000[4] = -1.87456362e-05L;
-      _NO2_coeffs_200_1000[5] =  9.14553637e-09L;
-      _NO2_coeffs_200_1000[6] = -1.77766146e-12L;
-      _NO2_coeffs_200_1000[7] = -1.54793043e+03L;
-      _NO2_coeffs_200_1000[8] =  4.06785541e+01L;
-    }
-
-    void init_NO2_coeffs_1000_6000()
-    {
-      _NO2_coeffs_1000_6000.resize(9);
-      _NO2_coeffs_1000_6000[0] =  7.21271176e+05L;
-      _NO2_coeffs_1000_6000[1] = -3.83253763e+03L;
-      _NO2_coeffs_1000_6000[2] =  1.11395534e+01L;
-      _NO2_coeffs_1000_6000[3] = -2.23801440e-03L;
-      _NO2_coeffs_1000_6000[4] =  6.54762164e-07L;
-      _NO2_coeffs_1000_6000[5] = -7.61120803e-11L;
-      _NO2_coeffs_1000_6000[6] =  3.32829926e-15L;
-      _NO2_coeffs_1000_6000[7] =  2.50244718e+04L;
-      _NO2_coeffs_1000_6000[8] = -4.30507224e+01L;
-    }
-
-    void init_all_NO2_coeffs()
-    {
-      _all_NO2_coeffs.insert(  _all_NO2_coeffs.end(),
-                              _NO2_coeffs_200_1000.begin(),
-                              _NO2_coeffs_200_1000.end() );
-
-      _all_NO2_coeffs.insert(  _all_NO2_coeffs.end(),
-                              _NO2_coeffs_1000_6000.begin(),
-                              _NO2_coeffs_1000_6000.end() );
-    }
-
-  protected:
-
-    std::vector<Scalar> _N2_coeffs_200_1000;
-    std::vector<Scalar> _N2_coeffs_1000_6000;
-    std::vector<Scalar> _N2_coeffs_6000_20000;
-
-    std::vector<Scalar> _NO2_coeffs_200_1000;
-    std::vector<Scalar> _NO2_coeffs_1000_6000;
-
-    std::vector<Scalar> _all_N2_coeffs;
-    std::vector<Scalar> _all_NO2_coeffs;
   };
 
   class NASA9CurveFitTestFloat : public NASA9CurveFitTest<float>
